requirements.txt
streamlit
tensorflow-cpu
opencv-python
pillow
mysql-connector-python
